+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -10,6 +10,7 @@
 	* Removed `ClaimList`
 	* Removed `ClaimListAccess`
 * Removed `Claims::getBestClaims` (you can use `StatementList::getBestStatements` instead)
+* Removed `Claims::getByRank` and `Claims::getByRanks` (you can use `StatementList::getWithRank` instead)
 
 ## Version 2.6.0 (2015-03-08)